AI/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
addict=2.4.0=pypi_0
blas=1.0=mkl
ca-certificates=2021.1.19=haa95532_0
certifi=2020.12.5=py38haa95532_0
click=7.1.2=pypi_0
cudatoolkit=10.2.89=h74a9793_1
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
flask=1.1.2=pypi_0
freetype=2.10.4=hd328e21_0
imageio=2.9.0=pypi_0
intel-openmp=2020.2=254
itsdangerous=1.1.0=pypi_0
jinja2=2.11.3=pypi_0
joblib=1.0.1=pypi_0
jpeg=9b=hb83a4c4_2
kiwisolver=1.3.1=pypi_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_1
libuv=1.40.0=he774522_0
lz4-c=1.9.3=h2bbff1b_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py38h196d8e1_0
mkl_fft=1.2.0=py38h45dec08_0
mkl_random=1.1.1=py38h47e9c7a_0
mmcv=1.2.6=pypi_0
mmfashion=0.4.0=pypi_0
networkx=2.5=pypi_0
ninja=1.10.2=py38h6d14046_0
numpy=1.19.2=py38hadc3359_0
numpy-base=1.19.2=py38ha3acd2a_0
olefile=0.46=py_0
opencv-python=4.5.1.48=pypi_0
openssl=1.1.1i=h2bbff1b_0
pillow=8.1.0=py38h4fa10fc_0
pip=20.3.3=py38haa95532_0
pyparsing=3.0.0b2=pypi_0
python=3.8.5=h5fd99cc_1
python-dateutil=2.8.1=pypi_0
pytorch=1.7.1=py3.8_cuda102_cudnn7_0
pywavelets=1.1.1=pypi_0
pyyaml=5.4.1=pypi_0
regex=2020.11.13=pypi_0
scikit-image=0.18.1=pypi_0
scikit-learn=0.24.1=pypi_0
scipy=1.6.0=pypi_0
setuptools=52.0.0=py38haa95532_0
six=1.15.0=py38haa95532_0
sqlite=3.33.0=h2a8f88b_0
threadpoolctl=2.1.0=pypi_0
tifffile=2021.2.1=pypi_0
tk=8.6.10=he774522_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cu102
typing_extensions=3.7.4.3=pyha847dfd_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py38_0
xz=5.2.5=h62dcd97_0
yapf=0.30.0=pypi_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
